角度のオフセット用変数の実体がエンコーダのカウント値であるためそれをfloatからuint_32に変えた

+++ Core/Inc/STM_MotorSystem.hpp
@@ -66,7 +66,7 @@ private:
 
 	float before_vel;
 
-	float ofset_angle;
+	uint32_t ofset_angle;
 	float before_angle;
 
 	GPIO_TypeDef *GPIO_dir;
